requirements.txt
langchain
langchain_groq
langchain_community
streamlit
python-dotenv
numpy
arxiv
wikipedia-api
duckduckgo-search
wikipedia